Seperated java activity/fragment files into Home, Login, and Setup folders
