+++ app/src/main/AndroidManifest.xml
@@ -13,16 +13,16 @@
         android:theme="@style/Theme.BadgerooProject"
         tools:targetApi="31">
         <activity
-            android:name=".FindRoommatesActivity"
+            android:name=".Home.FindRoommatesActivity"
             android:exported="false" />
         <activity
-            android:name=".AccountSetupActivity"
+            android:name=".Setup.AccountSetupActivity"
             android:exported="false" />
         <activity
-            android:name=".RegisterActivity"
+            android:name=".Setup.RegisterActivity"
             android:exported="false" />
         <activity
-            android:name=".LoginActivity"
+            android:name=".Login.LoginActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
