+++ app/src/main/java/com/cs407/badgerooproject/Login/LoginActivity.java
@@ -1,8 +1,10 @@
-package com.cs407.badgerooproject;
+package com.cs407.badgerooproject.Login;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
+
+import com.cs407.badgerooproject.R;
 
 // Fragments: ForgotPasswordFragment, ChangePasswordFragment
