+++ app/src/main/java/com/cs407/badgerooproject/Setup/UploadProfilePictureFragment.java
@@ -1,20 +1,23 @@
-package com.cs407.badgerooproject;
+package com.cs407.badgerooproject.Setup;
 
 import android.os.Bundle;
+
 import androidx.fragment.app.Fragment;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-// Activity: FindRoommatesActivity
+import com.cs407.badgerooproject.R;
+
+// Activity: AccountSetupActivity
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link SettingsFragment#newInstance} factory method to
+ * Use the {@link UploadProfilePictureFragment#newInstance} factory method to
  * create an instance of this fragment.
- *
  */
-public class SettingsFragment extends Fragment {
+public class UploadProfilePictureFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -25,26 +28,26 @@ public class SettingsFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
+    public UploadProfilePictureFragment() {
+        // Required empty public constructor
+    }
+
     /**
      * Use this factory method to create a new instance of
      * this fragment using the provided parameters.
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment SettingsFragment.
+     * @return A new instance of fragment UploadProfilePictureFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static SettingsFragment newInstance(String param1, String param2) {
-        SettingsFragment fragment = new SettingsFragment();
+    public static UploadProfilePictureFragment newInstance(String param1, String param2) {
+        UploadProfilePictureFragment fragment = new UploadProfilePictureFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
         fragment.setArguments(args);
         return fragment;
-    }
-
-    public SettingsFragment() {
-        // Required empty public constructor
     }
 
     @Override
@@ -60,6 +63,6 @@ public class SettingsFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_settings, container, false);
+        return inflater.inflate(R.layout.fragment_upload_profile_picture, container, false);
     }
 }
