+++ app/src/main/res/layout/activity_account_setup.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AccountSetupActivity">
+    tools:context=".Setup.AccountSetupActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
