+++ app/src/main/res/layout/activity_find_roommates.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FindRoommatesActivity">
+    tools:context=".Home.FindRoommatesActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
