+++ app/src/main/res/layout/fragment_messages_list.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MessagesListFragment">
+    tools:context=".Home.MessagesListFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
