+++ app/src/main/res/layout/fragment_setup_preferences.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SetupPreferencesFragment">
+    tools:context=".Setup.SetupPreferencesFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
