+++ app/src/main/res/layout/fragment_upload_profile_picture.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UploadProfilePictureFragment">
+    tools:context=".Setup.UploadProfilePictureFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
